test: improve coverage for validator service
