+++ validator/tests/test_validator_services.py
@@ -1,0 +1,555 @@
+import uuid
+from django.test import TestCase
+from unittest.mock import patch, Mock
+from requests.exceptions import RequestException
+from validator.constants import FeedbackMsg, ErrorMsg
+from validator.enums import ValidationType
+from validator.exceptions import AIServiceErrorException
+from cause.models import Causes
+from question.models import Question
+from validator.services import CausesService
+
+
+class CausesServiceTest(TestCase):
+    @patch('validator.services.Groq')
+    def test_api_call_positive(self, mock_groq):
+        mock_client = Mock()
+        mock_chat_completion = Mock()
+        mock_chat_completion.choices = [Mock(message=Mock(content='true'))]
+        mock_client.chat.completions.create.return_value = mock_chat_completion
+        mock_groq.return_value = mock_client
+
+        service = CausesService()
+        system_message = "You are an AI model. You are asked to determine whether the given cause is the cause of the given problem."
+        user_prompt = "Is 'Example cause' the cause of 'Example problem'? Answer only with True/False"
+        response = service.api_call(system_message, user_prompt, ValidationType.NORMAL)
+
+        self.assertEqual(response, 1)
+        mock_client.chat.completions.create.assert_called_once_with(
+            messages=[
+                {
+                    "role": "system",
+                    "content": system_message,
+                },
+                {
+                    "role": "user",
+                    "content": user_prompt
+                }
+            ],
+            model="deepseek-r1-distill-llama-70b",
+            temperature=0.7,
+            max_completion_tokens=1024,
+            top_p=0.95,
+            stream=False,
+            reasoning_format="hidden",
+            seed=42
+        )
+
+    @patch('validator.services.Groq')
+    def test_api_call_returns_false(self, mock_groq):
+        mock_client = Mock()
+        mock_chat_completion = Mock()
+        mock_chat_completion.choices = [Mock(message=Mock(content='false'))]
+        mock_client.chat.completions.create.return_value = mock_chat_completion
+        mock_groq.return_value = mock_client
+
+        service = CausesService()
+        system_message = "You are an AI model. You are asked to determine whether the given cause is the cause of the given problem."
+        user_prompt = "Is 'Example cause' the cause of 'Example problem'? Answer only with True/False"
+        response = service.api_call(system_message, user_prompt, ValidationType.NORMAL)
+
+        self.assertEqual(response, 0)
+
+    @patch('validator.services.Groq')
+    def test_api_call_request_exception(self, mock_groq):
+        mock_client = Mock()
+        mock_client.chat.completions.create.side_effect = RequestException("Network error")
+        mock_groq.return_value = mock_client
+
+        service = CausesService()
+        system_message = "You are an AI model. You are asked to determine whether the given cause is the cause of the given problem."
+        user_prompt = "Is 'Example cause' the cause of 'Example problem'? Answer only with True/False"
+
+        with self.assertRaises(AIServiceErrorException) as context:
+            service.api_call(system_message, user_prompt, ValidationType.NORMAL)
+
+        mock_client.chat.completions.create.assert_called_once_with(
+            messages=[
+                {
+                    "role": "system",
+                    "content": system_message,
+                },
+                {
+                    "role": "user",
+                    "content": user_prompt
+                }
+            ],
+            model="deepseek-r1-distill-llama-70b",
+            temperature=0.7,
+            max_completion_tokens=1024,
+            top_p=0.95,
+            stream=False,
+            reasoning_format="hidden",
+            seed=42
+        )
+
+        self.assertEqual(str(context.exception), "Failed to call the AI service.")
+
+    @patch('validator.services.Groq')
+    def test_api_call_negative(self, mock_groq):
+        mock_client = Mock()
+        mock_client.chat.completions.create.side_effect = Exception("API call failed")
+        mock_groq.return_value = mock_client
+
+        service = CausesService()
+        system_message = "You are an AI model. You are asked to determine whether the given cause is the cause of the given problem."
+        user_prompt = "test prompt"
+        with self.assertRaises(Exception):
+            service.api_call(system_message, user_prompt, ValidationType.NORMAL)
+
+    @patch('validator.services.Groq')
+    def test_api_call_forbidden_access(self, mock_groq):
+        mock_client = Mock()
+        mock_client.chat.completions.create.side_effect = Exception("Unauthorized: Invalid API key")
+        mock_groq.return_value = mock_client
+
+        service = CausesService()
+        system_message = "You are an AI model. You are asked to determine whether the given cause is the cause of the given problem."
+        user_prompt = "Test prompt"
+        with self.assertRaises(Exception):
+            service.api_call(system_message, user_prompt, ValidationType.NORMAL)
+
+    @patch('validator.services.Groq')
+    def test_api_call_validation_type_root(self, mock_groq):
+        mock_client = Mock()
+        mock_chat_completion = Mock()
+        mock_chat_completion.choices = [Mock(message=Mock(content='true'))]
+        mock_client.chat.completions.create.return_value = mock_chat_completion
+        mock_groq.return_value = mock_client
+
+        service = CausesService()
+        response = service.api_call(
+            system_message="test",
+            user_prompt="test",
+            validation_type=ValidationType.ROOT
+        )
+        self.assertEqual(response, 1)
+
+    @patch('validator.services.Groq')
+    def test_api_call_validation_type_root_type(self, mock_groq):
+        mock_client = Mock()
+        mock_chat_completion = Mock()
+        mock_chat_completion.choices = [Mock(message=Mock(content='2'))]
+        mock_client.chat.completions.create.return_value = mock_chat_completion
+        mock_groq.return_value = mock_client
+
+        service = CausesService()
+        response = service.api_call(
+            system_message="test",
+            user_prompt="test",
+            validation_type=ValidationType.ROOT_TYPE
+        )
+        self.assertEqual(response, 2)
+
+    @patch('validator.services.Groq')
+    def test_retrieve_feedback_not_cause_1_row(self, mock_groq):
+        mock_client = Mock()
+        mock_chat_completion = Mock()
+        mock_chat_completion.choices = [Mock(message=Mock(content='1'))]
+        mock_client.chat.completions.create.return_value = mock_chat_completion
+        mock_groq.return_value = mock_client
+
+        service = CausesService()
+        cause = Causes(question_id=uuid.uuid4(), cause="False Cause", row=1, column=1)
+        problem = Question(question="Test problem")
+
+        service.retrieve_feedback(cause, problem, None)
+        self.assertEqual(cause.feedback, FeedbackMsg.FALSE_ROW_1_NOT_CAUSE.format(column='B'))
+
+    @patch('validator.services.Groq')
+    def test_retrieve_feedback_positive_neutral_1_row(self, mock_groq):
+        mock_client = Mock()
+        mock_chat_completion = Mock()
+        mock_chat_completion.choices = [Mock(message=Mock(content='2'))]
+        mock_client.chat.completions.create.return_value = mock_chat_completion
+        mock_groq.return_value = mock_client
+
+        service = CausesService()
+        cause = Causes(question_id=uuid.uuid4(), cause="Positive/Neutral Cause", row=1, column=1)
+        problem = Question(question="Test problem")
+
+        service.retrieve_feedback(cause, problem, None)
+        self.assertEqual(cause.feedback, FeedbackMsg.FALSE_ROW_N_POSITIVE_NEUTRAL.format(column='B', row=1))
+
+    @patch('validator.services.Groq')
+    def test_retrieve_feedback_not_cause_n_row(self, mock_groq):
+        mock_client = Mock()
+        mock_chat_completion = Mock()
+        mock_chat_completion.choices = [Mock(message=Mock(content='1'))]
+        mock_client.chat.completions.create.return_value = mock_chat_completion
+        mock_groq.return_value = mock_client
+
+        service = CausesService()
+        prev_cause = Causes(question_id=uuid.uuid4(), cause="Base Cause", row=1, column=1)
+        cause = Causes(question_id=uuid.uuid4(), cause="False Cause", row=2, column=1)
+        problem = Question(question="Test problem")
+
+        service.retrieve_feedback(cause, problem, prev_cause)
+        self.assertEqual(cause.feedback, FeedbackMsg.FALSE_ROW_N_NOT_CAUSE.format(column='B', row=2, prev_row=1))
+
+    @patch('validator.services.Groq')
+    def test_retrieve_feedback_positive_neutral_n_row(self, mock_groq):
+        mock_client = Mock()
+        mock_chat_completion = Mock()
+        mock_chat_completion.choices = [Mock(message=Mock(content='2'))]
+        mock_client.chat.completions.create.return_value = mock_chat_completion
+        mock_groq.return_value = mock_client
+
+        service = CausesService()
+        prev_cause = Causes(question_id=uuid.uuid4(), cause="Base Cause", row=1, column=1)
+        cause = Causes(question_id=uuid.uuid4(), cause="Positive/Neutral Cause", row=2, column=1)
+        problem = Question(question="Test problem")
+
+        service.retrieve_feedback(cause, problem, prev_cause)
+        self.assertEqual(cause.feedback, FeedbackMsg.FALSE_ROW_N_POSITIVE_NEUTRAL.format(column='B', row=2))
+
+    @patch('validator.services.Groq')
+    def test_retrieve_feedback_similar_previous_n_row(self, mock_groq):
+        mock_client = Mock()
+        mock_chat_completion = Mock()
+        mock_chat_completion.choices = [Mock(message=Mock(content='3'))]
+        mock_client.chat.completions.create.return_value = mock_chat_completion
+        mock_groq.return_value = mock_client
+
+        service = CausesService()
+        cause = Causes(question_id=uuid.uuid4(), cause="Similar Cause", row=2, column=1)
+        prev_cause = Causes(question_id=uuid.uuid4(), cause="Previous Cause", row=1, column=1)
+        problem = Question(question="Test problem")
+
+        service.retrieve_feedback(cause, problem, prev_cause)
+        self.assertEqual(cause.feedback, FeedbackMsg.FALSE_ROW_N_SIMILAR_PREVIOUS.format(column='B', row=2))
+
+    @patch('validator.services.Groq')
+    def test_retrieve_feedback_with_prev_cause(self, mock_groq):
+        mock_client = Mock()
+        mock_chat_completion = Mock()
+        mock_chat_completion.choices = [Mock(message=Mock(content='1'))]
+        mock_client.chat.completions.create.return_value = mock_chat_completion
+        mock_groq.return_value = mock_client
+
+        service = CausesService()
+        question_id = uuid.uuid4()
+        question = Question.objects.create(
+            id=question_id,
+            question="Test question"
+        )
+        cause = Causes.objects.create(
+            question_id=question_id,
+            cause="Test cause",
+            row=1,
+            column=0,
+            status=False
+        )
+        prev_cause = Causes.objects.create(
+            question_id=question_id,
+            cause="Previous cause",
+            row=0,
+            column=0
+        )
+
+        service.retrieve_feedback(cause, question, prev_cause)
+        self.assertIn("bukan merupakan sebab", cause.feedback)
+
+    @patch('validator.services.Groq')
+    def test_retrieve_feedback_without_prev_cause(self, mock_groq):
+        mock_client = Mock()
+        mock_chat_completion = Mock()
+        mock_chat_completion.choices = [Mock(message=Mock(content='2'))]
+        mock_client.chat.completions.create.return_value = mock_chat_completion
+        mock_groq.return_value = mock_client
+
+        service = CausesService()
+        question_id = uuid.uuid4()
+        question = Question.objects.create(
+            id=question_id,
+            question="Test question"
+        )
+        cause = Causes.objects.create(
+            question_id=question_id,
+            cause="Test cause",
+            row=1,
+            column=0,
+            status=False
+        )
+
+        service.retrieve_feedback(cause, question, None)
+        self.assertIn("sebab positif atau netral", cause.feedback)
+
+    @patch('validator.services.Groq')
+    def test_validate_with_status_true(self, mock_groq):
+        service = CausesService()
+        question_id = uuid.uuid4()
+        question = Question.objects.create(
+            id=question_id,
+            question="Test question"
+        )
+        cause = Causes.objects.create(
+            question_id=question_id,
+            cause="Test cause",
+            row=1,
+            column=0,
+            status=True
+        )
+        
+        service.validate(question_id)
+        self.assertTrue(cause.status)
+
+    @patch('validator.services.Groq')
+    def test_validate_with_multiple_rows(self, mock_groq):
+        mock_client = Mock()
+        mock_chat_completion = Mock()
+        mock_chat_completion.choices = [Mock(message=Mock(content='true'))]
+        mock_client.chat.completions.create.return_value = mock_chat_completion
+        mock_groq.return_value = mock_client
+
+        service = CausesService()
+        question_id = uuid.uuid4()
+        question = Question.objects.create(
+            id=question_id,
+            question="Test question"
+        )
+
+        first_cause = Causes.objects.create(
+            question_id=question_id,
+            cause="First cause",
+            row=1,
+            column=0,
+            status=True
+        )
+
+        cause = Causes.objects.create(
+            question_id=question_id,
+            cause="Test cause",
+            row=2,
+            column=0,
+            status=False
+        )
+
+        def mock_api_call_side_effect(self, system_message, user_prompt, validation_type):
+            if validation_type == ValidationType.NORMAL:
+                return 1
+            elif validation_type == ValidationType.ROOT:
+                return 0
+            return 0
+
+        with patch.object(CausesService, 'api_call', side_effect=mock_api_call_side_effect) as mock_api_call:
+            service.validate(question_id)
+            cause.refresh_from_db()
+            self.assertTrue(cause.status)
+            
+            mock_api_call.assert_any_call(
+                self=service,
+                system_message="You are an AI model. You are asked to determine whether the given cause is the cause of the given problem.",
+                user_prompt=f"Is '{cause.cause}' the cause of '{first_cause.cause}'? Answer only with True/False",
+                validation_type=ValidationType.NORMAL
+            )
+
+    @patch('validator.services.Groq')
+    def test_check_root_cause_with_korupsi_categories(self, mock_groq):
+        mock_client = Mock()
+        mock_chat_completion = Mock()
+        mock_chat_completion.choices = [Mock(message=Mock(content='true'))]
+        mock_client.chat.completions.create.side_effect = [
+            Mock(choices=[Mock(message=Mock(content='true'))]),
+            Mock(choices=[Mock(message=Mock(content='2'))])
+        ]
+        mock_client.chat.completions.create.return_value = mock_chat_completion
+        mock_groq.return_value = mock_client
+
+        service = CausesService()
+        question_id = uuid.uuid4()
+        question = Question.objects.create(
+            id=question_id,
+            question="Test question"
+        )
+        cause = Causes.objects.create(
+            question_id=question_id,
+            cause="Test cause",
+            row=1,
+            column=0,
+            status=False
+        )
+
+        service.check_root_cause(cause, question)
+        cause.refresh_from_db()
+        self.assertTrue(cause.root_status)
+        self.assertIn("Korupsi Tahta", cause.feedback)
+
+    @patch('validator.services.Groq')
+    def test_check_root_cause_default_korupsi(self, mock_groq):
+        mock_client = Mock()
+        mock_chat_completion = Mock()
+        mock_chat_completion.choices = [Mock(message=Mock(content='true'))]
+        mock_client.chat.completions.create.side_effect = [
+            Mock(choices=[Mock(message=Mock(content='true'))]),
+            Mock(choices=[Mock(message=Mock(content='4'))])
+        ]
+        mock_client.chat.completions.create.return_value = mock_chat_completion
+        mock_groq.return_value = mock_client
+
+        service = CausesService()
+        question_id = uuid.uuid4()
+        question = Question.objects.create(
+            id=question_id,
+            question="Test question"
+        )
+        cause = Causes.objects.create(
+            question_id=question_id,
+            cause="Test cause",
+            row=1,
+            column=0,
+            status=False
+        )
+
+        service.check_root_cause(cause, question)
+        cause.refresh_from_db()
+        self.assertTrue(cause.root_status)
+        self.assertIn("Korupsi Harta", cause.feedback)
+
+    @patch('validator.services.Groq')
+    def test_validate_with_row_1(self, mock_groq):
+        mock_client = Mock()
+        mock_chat_completion = Mock()
+        mock_chat_completion.choices = [Mock(message=Mock(content='true'))]
+        mock_client.chat.completions.create.return_value = mock_chat_completion
+        mock_groq.return_value = mock_client
+
+        service = CausesService()
+        question_id = uuid.uuid4()
+        question = Question.objects.create(
+            id=question_id,
+            question="Test question"
+        )
+        cause = Causes.objects.create(
+            question_id=question_id,
+            cause="Test cause",
+            row=1,
+            column=0,
+            status=False
+        )
+
+        def mock_api_call_side_effect(self, system_message, user_prompt, validation_type):
+            if validation_type == ValidationType.NORMAL:
+                return 1
+            return 0
+
+        with patch.object(CausesService, 'api_call', side_effect=mock_api_call_side_effect) as mock_api_call:
+            service.validate(question_id)
+            cause.refresh_from_db()
+            self.assertTrue(cause.status)
+            
+            mock_api_call.assert_any_call(
+                self=service,
+                system_message="You are an AI model. You are asked to determine whether the given cause is the cause of the given problem.",
+                user_prompt=f"Is '{cause.cause}' the cause of this question: '{question.question}'? Answer only with True/False",
+                validation_type=ValidationType.NORMAL
+            )
+
+    @patch('validator.services.Groq')
+    def test_validate_returns_false(self, mock_groq):
+        mock_client = Mock()
+        mock_chat_completion = Mock()
+        mock_chat_completion.choices = [Mock(message=Mock(content='false'))]
+        mock_client.chat.completions.create.return_value = mock_chat_completion
+        mock_groq.return_value = mock_client
+
+        service = CausesService()
+        question_id = uuid.uuid4()
+        question = Question.objects.create(
+            id=question_id,
+            question="Test question"
+        )
+        first_cause = Causes.objects.create(
+            question_id=question_id,
+            cause="First cause",
+            row=1,
+            column=0,
+            status=True
+        )
+        cause = Causes.objects.create(
+            question_id=question_id,
+            cause="Test cause",
+            row=2,
+            column=0,
+            status=False
+        )
+
+        def mock_api_call_side_effect(self, system_message, user_prompt, validation_type):
+            return 0
+
+        with patch.object(CausesService, 'api_call', side_effect=mock_api_call_side_effect) as mock_api_call:
+            service.validate(question_id)
+            cause.refresh_from_db()
+            self.assertFalse(cause.status)
+
+    @patch('validator.services.Groq')
+    def test_check_root_cause_with_korupsi_harta(self, mock_groq):
+        mock_client = Mock()
+        mock_chat_completion = Mock()
+        mock_chat_completion.choices = [Mock(message=Mock(content='true'))]
+        mock_client.chat.completions.create.side_effect = [
+            Mock(choices=[Mock(message=Mock(content='true'))]),
+            Mock(choices=[Mock(message=Mock(content='1'))])
+        ]
+        mock_client.chat.completions.create.return_value = mock_chat_completion
+        mock_groq.return_value = mock_client
+
+        service = CausesService()
+        question_id = uuid.uuid4()
+        question = Question.objects.create(
+            id=question_id,
+            question="Test question"
+        )
+        cause = Causes.objects.create(
+            question_id=question_id,
+            cause="Test cause",
+            row=1,
+            column=0,
+            status=False
+        )
+
+        service.check_root_cause(cause, question)
+        cause.refresh_from_db()
+        self.assertTrue(cause.root_status)
+        self.assertIn("Korupsi Harta", cause.feedback)
+
+    @patch('validator.services.Groq')
+    def test_check_root_cause_with_korupsi_cinta(self, mock_groq):
+        mock_client = Mock()
+        mock_chat_completion = Mock()
+        mock_chat_completion.choices = [Mock(message=Mock(content='true'))]
+        mock_client.chat.completions.create.side_effect = [
+            Mock(choices=[Mock(message=Mock(content='true'))]),
+            Mock(choices=[Mock(message=Mock(content='3'))])
+        ]
+        mock_client.chat.completions.create.return_value = mock_chat_completion
+        mock_groq.return_value = mock_client
+
+        service = CausesService()
+        question_id = uuid.uuid4()
+        question = Question.objects.create(
+            id=question_id,
+            question="Test question"
+        )
+        cause = Causes.objects.create(
+            question_id=question_id,
+            cause="Test cause",
+            row=1,
+            column=0,
+            status=False
+        )
+
+        service.check_root_cause(cause, question)
+        cause.refresh_from_db()
+        self.assertTrue(cause.root_status)
+        self.assertIn("Korupsi Cinta", cause.feedback)
